+++ src/Row.tsx
@@ -33,6 +33,7 @@ class Row extends React.PureComponent<RowProps> {
       headerHeight,
       rowRef,
       classPrefix,
+      children,
       ...rest
     } = this.props;
 
@@ -53,7 +54,11 @@ class Row extends React.PureComponent<RowProps> {
       <TableContext.Consumer>
         {({ translateDOMPositionXY }) => {
           translateDOMPositionXY?.(styles, 0, top);
-          return <div {...unhandledProps} ref={rowRef} className={classes} style={styles} />;
+          return (
+            <div role="row" {...unhandledProps} ref={rowRef} className={classes} style={styles}>
+              {children}
+            </div>
+          );
         }}
       </TableContext.Consumer>
     );