+++ test/CellSpec.js
@@ -3,6 +3,7 @@ import ReactTestUtils from 'react-dom/test-utils';
 
 import { getDOMNode } from './TestWrapper';
 import Cell from '../src/Cell';
+import TableContext from '../src/TableContext';
 import { LAYER_WIDTH } from '../src/constants';
 
 describe('Cell', () => {
@@ -85,8 +86,25 @@ describe('Cell', () => {
   });
 
   it('Should have a expand icon', () => {
-    const instanceDom = getDOMNode(<Cell hasChildren firstColumn />);
+    const instanceDom = getDOMNode(
+      <div>
+        <TableContext.Provider value={{ isTree: true }}>
+          <Cell hasChildren firstColumn />
+        </TableContext.Provider>
+      </div>
+    );
     assert.ok(instanceDom.querySelector('.rs-table-cell-expand-icon'));
+  });
+
+  it('Should have a expanded icon', () => {
+    const instanceDom = getDOMNode(
+      <div>
+        <TableContext.Provider value={{ isTree: true }}>
+          <Cell hasChildren firstColumn expanded />
+        </TableContext.Provider>
+      </div>
+    );
+    assert.ok(instanceDom.querySelector('[aria-label="arrow-down"]'));
   });
 
   it('Should be 60 the left', () => {
@@ -104,14 +122,18 @@ describe('Cell', () => {
     };
 
     const instanceDom = getDOMNode(
-      <Cell
-        hasChildren
-        firstColumn
-        onTreeToggle={doneOp}
-        rowData={{ name: 'a' }}
-        rowIndex={1}
-        rowKey="name"
-      />
+      <div>
+        <TableContext.Provider value={{ isTree: true }}>
+          <Cell
+            hasChildren
+            firstColumn
+            onTreeToggle={doneOp}
+            rowData={{ name: 'a' }}
+            rowIndex={1}
+            rowKey="name"
+          />
+        </TableContext.Provider>
+      </div>
     );
 
     ReactTestUtils.Simulate.click(instanceDom.querySelector('.rs-table-cell-expand-icon'));